# docs/openapi.yaml
openapi: 3.0.0
info:
  title: CTI Platform API
  version: 0.1.0
  description: Cyber Threat Intelligence Platform REST API
  contact:
    name: CTI Team
    email: support@cti-platform.local

servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://api.cti-platform.com
    description: Production

tags:
  - name: articles
    description: Article management and retrieval
  - name: sources
    description: RSS source configuration
  - name: search
    description: Search and facets

paths:
  /v1/articles:
    get:
      tags: [articles]
      summary: Get articles with filters
      parameters:
        - in: query
          name: query
          schema:
            type: string
          description: Full-text search query
        - in: query
          name: from
          schema:
            type: string
            format: date-time
          description: Start date (ISO 8601)
        - in: query
          name: to
          schema:
            type: string
            format: date-time
          description: End date (ISO 8601)
        - in: query
          name: type
          schema:
            type: string
          description: Source type filter
        - in: query
          name: source
          schema:
            type: string
            format: uuid
          description: Source ID filter
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: size
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: List of articles
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Article'
                  total:
                    type: integer
                  page:
                    type: integer
                  size:
                    type: integer

  /v1/articles/{id}:
    get:
      tags: [articles]
      summary: Get article by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Article details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleDetail'
        '404':
          description: Article not found

  /v1/articles/{id}/click:
    post:
      tags: [articles]
      summary: Track click and return source URL
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Source URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  sourceUrl:
                    type: string
                    format: uri
        '404':
          description: Article not found

  /v1/sources:
    get:
      tags: [sources]
      summary: List all RSS sources
      responses:
        '200':
          description: List of sources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Source'
    
    post:
      tags: [sources]
      summary: Add new RSS source
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSource'
      responses:
        '201':
          description: Source created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'

  /v1/search/fields:
    get:
      tags: [search]
      summary: Get search facets and aggregations
      responses:
        '200':
          description: Facets data
          content:
            application/json:
              schema:
                type: object

components:
  schemas:
    Article:
      type: object
      properties:
        id:
          type: string
          format: uuid
        titleEs:
          type: string
        summaryEs:
          type: string
        tags:
          type: array
          items:
            type: string
        iocsPreview:
          type: array
          items:
            type: string
        publishedAt:
          type: string
          format: date-time
        sourceName:
          type: string
        translated:
          type: boolean
        confidence:
          type: number
          format: float
        hasIocs:
          type: boolean
        iocCount:
          type: integer

    ArticleDetail:
      allOf:
        - $ref: '#/components/schemas/Article'
        - type: object
          properties:
            bodyEs:
              type: string
            sourceUrl:
              type: string
              format: uri
            languageDetected:
              type: string
            iocs:
              type: array
              items:
                $ref: '#/components/schemas/IOC'

    IOC:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [ip, domain, url, hash_md5, hash_sha1, hash_sha256, cve, email]
        value:
          type: string
        confidence:
          type: number
          format: float
        context:
          type: string

    Source:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        url:
          type: string
          format: uri
        type:
          type: string
        region:
          type: string
        country:
          type: string
        language:
          type: string
        enabled:
          type: boolean
        fetchIntervalMinutes:
          type: integer
        lastFetchedAt:
          type: string
          format: date-time

    CreateSource:
      type: object
      required:
        - name
        - url
      properties:
        name:
          type: string
        url:
          type: string
          format: uri
        type:
          type: string
        region:
          type: string
        country:
          type: string
        language:
          type: string
        enabled:
          type: boolean
        fetchIntervalMinutes:
          type: integer